Refactor : RectangleModel를 Plane으로 변경 Plane 요구 사항 (터치 좌표 확인)을 보고 plane에서 이벤트를 발생시키는 게 맞는 거 같아 설계 변경
