+++ app/src/main/java/com/codesquard/kotlin_drawingapp/TaskPresenter.kt
@@ -1,13 +1,23 @@
 package com.codesquard.kotlin_drawingapp
 
-class TaskPresenter(val taskView: TaskContract.TaskView) : TaskContract.Presenter {
-    private val plane = Plane()
+class TaskPresenter(val taskView: TaskContract.TaskView) : TaskContract.Presenter,
+    RectangleListener {
+
+    private val rectangleList = mutableListOf<Rectangle>()
 
     override fun addNewRectangle() {
-       val newRect = plane.getNewRect()
+        val newRect = getNewRect()
         taskView.showRectangle(newRect)
     }
 
+    override fun onCreateRectangle(newRect: Rectangle) {
+        rectangleList.add(newRect)
+    }
+
+    private fun getNewRect(): Rectangle {
+        Plane(this).createNewRectangle()
+        return rectangleList[rectangleList.size - 1]
+    }
 
 }
